requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary>=2.9.9
pyodbc==5.0.1
APScheduler==3.10.4
python-decouple==3.8
django-filter==23.5
Pillow>=10.3.0
django-extensions==3.2.3
